requirements.txt
python-telegram-bot==13.15
httpx~=0.24.1
environs==5.2.1
